Added xp system

+++ src/main/java/io/github/thelordman/costrength/utilities/Data.java
@@ -1,12 +1,10 @@
 package io.github.thelordman.costrength.utilities;
 
-import org.bukkit.boss.BossBar;
 import org.bukkit.entity.Player;
-import org.javatuples.Pair;
 
 import java.util.HashMap;
 
 public class Data {
-    public static HashMap<Player, Integer> combatTag = new HashMap<>();
+    public static HashMap<Player, Byte> combatTag = new HashMap<>();
     public static HashMap<Player, Float> bounty = new HashMap<>();
 }
